Add files via upload
This is an implementation of the Pathfinding A* Algorithm, which uses a smart guess (a heuristic estimation) to find the shortest path from a certain point to a destination point in a given maze. This algorithm is very useful and can be implemented and is being used for different goals, such as finding shortest way from point A to B in a map for navigation purposes for example.